+++ LocationStation/LocationStation/ContentView.swift
@@ -9,10 +9,9 @@ import SwiftUI
 
 struct ContentView: View {
     @StateObject private var viewModel = ContentViewModel()
-    @State private var region = MKCoordinateRegion(center: CLLocationCoordinate2D(latitude: 50.92897235052956, longitude: 6.943178858308985), span: MKCoordinateSpan(latitudeDelta: 0.1, longitudeDelta: 0.1))
     
     var body: some View {
-        Map(coordinateRegion: $region, showsUserLocation: true)
+        Map(coordinateRegion: $viewModel.region, showsUserLocation: true)
             .ignoresSafeArea()
             .onAppear{
                 viewModel.checkIsLocationServiceOn()
@@ -23,40 +22,5 @@ struct ContentView: View {
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
         ContentView()
-    }
-}
-
-
-final class ContentViewModel: NSObject, ObservableObject, CLLocationManagerDelegate {
-    var locationManager: CLLocationManager?
-    
-    func checkIsLocationServiceOn(){
-        if CLLocationManager.locationServicesEnabled() {
-            locationManager = CLLocationManager()
-            locationManager!.delegate = self
-        } else {
-            // Todo show alert to tell user to turn location on
-        }
-    }
-    
-   private func handleLocationPermission() {
-        guard let locationManager = locationManager else { return }
-        switch locationManager.authorizationStatus {
-        
-        case .notDetermined:
-            locationManager.requestWhenInUseAuthorization()
-        case .restricted:
-           print("Your location is restricted, maybe check for parental controls.")
-        case .denied:
-            print("It seems as if you have denied location permission for this app, please go to settings and give location permissions to this app")
-        case .authorizedAlways, .authorizedWhenInUse:
-            break
-        @unknown default:
-            break
-        }
-    }
-    
-    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
-        handleLocationPermission()
     }
 }
